added all CSRF tokens

+++ public_html/News/storyDel.php
@@ -4,6 +4,7 @@ include 'includes/header.php';
 if(!($_SESSION["LoggedIn"])){
     header("Location: login.php");
 }
+
 ?>
 
 <!DOCTYPE html>
@@ -20,22 +21,36 @@ if(!($_SESSION["LoggedIn"])){
 
 <?php 
 
-echo ($_GET['articleID']);
-$articleID = $_GET['articleID'];
-$userName = $_SESSION['currUser'];
+$authToken = $_POST['authToken'];
 
-$stmt = $mysqli->prepare("delete from articles where article_id= ?");
-$stmt->bind_param('i', $articleID);
+if(!hash_equals($_SESSION['token'], $authToken)){
+    die("Warning: someone tried to forge a request");
+} else{
+    echo ($_POST['articleID']);
+    $articleID = $_POST['articleID'];
+    $userName = $_SESSION['currUser'];
+    
+    $stmt2 = $mysqli->prepare("delete from comments where article_id= ?");
+    $stmt2->bind_param('i', $articleID);
 
-if(!$stmt){
-	printf("Query Prep Failed: %s<br>", $mysqli->error);
-	exit;
-}else{
-    header("Location: profile.php");
+    $stmt2->execute();
+    $stmt2->close();
+
+    $stmt = $mysqli->prepare("delete from articles where article_id= ?");
+    $stmt->bind_param('i', $articleID);
+    
+    if(!$stmt){
+        printf("Query Prep Failed: %s<br>", $mysqli->error);
+        exit;
+    }else{
+        header("Location: profile.php");
+    }
+    
+    $stmt->execute();
+    $stmt->close();
+
 }
 
-$stmt->execute();
-$stmt->close();
